requirements.txt
beautifulsoup4==4.11.1
Django==4.1
djangorestframework==3.13.1
requests==2.26.0
toml==0.10.2
whitenoise==6.2.0
django-heroku==0.3.1
gunicorn==20.1.0
python-dotenv==0.20.0